warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6595 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfs.cxx
@@ -2802,7 +2802,7 @@ static void playerKilled(int victimIndex, int killerIndex, int reason,
 
   //update tk-score
   if ((victimIndex != killerIndex) && teamkill) {
-    bool isTk = killer->setAndTestTK(clOptions->teamKillerKickRatio);
+    bool isTk = killer->setAndTestTK((float)clOptions->teamKillerKickRatio);
     if (isTk) {
        char message[MessageLen];
        strcpy(message, "You have been automatically kicked for team killing" );
